dont remove users from db on leave if muted

+++ src/main/java/com/github/kaktushose/nplaybot/rank/JoinLeaveListener.java
@@ -1,5 +1,7 @@
 package com.github.kaktushose.nplaybot.rank;
 
+import com.github.kaktushose.nplaybot.Database;
+import com.github.kaktushose.nplaybot.permissions.PermissionsService;
 import net.dv8tion.jda.api.events.guild.member.GuildMemberJoinEvent;
 import net.dv8tion.jda.api.events.guild.member.GuildMemberRemoveEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
@@ -8,9 +10,11 @@ import org.jetbrains.annotations.NotNull;
 public class JoinLeaveListener extends ListenerAdapter {
 
     private final RankService rankService;
+    private final PermissionsService permissionsService;
 
-    public JoinLeaveListener(RankService rankService) {
-        this.rankService = rankService;
+    public JoinLeaveListener(Database database) {
+        this.rankService = database.getRankService();
+        this.permissionsService = database.getPermissionsService();
     }
 
     @Override
@@ -21,6 +25,9 @@ public class JoinLeaveListener extends ListenerAdapter {
 
     @Override
     public void onGuildMemberRemove(@NotNull GuildMemberRemoveEvent event) {
+        if (!permissionsService.hasUserPermissions(event.getMember())) {
+            return;
+        }
         rankService.removeUser(event.getUser());
     }
 }
